Use long unsigned int for iteration variable to avoid type comparision complaint
----

+++ compiler/resolution/wrappers.cpp
@@ -2234,7 +2234,7 @@ PromotionInfo::PromotionInfo(FnSymbol* fn,
       bool actualProvidedForFormal = false;
       // Can't use for_vector here as it will not iterate if the vector starts
       // with a NULL entry
-      for (int k = 0; k < actualFormals.size(); k++) {
+      for (long unsigned int k = 0; k < actualFormals.size(); k++) {
         ArgSymbol* arg = actualFormals[k];
         if (arg == NULL)
           continue;
